requirements.txt
fastapi==0.95.1
pydantic==1.10.2
uvicorn==0.22.0
openai==0.27.4
python-dotenv==1.0.0
langchain
langchain_community

